+++ lib/indexd_stack_dev.dart
@@ -1,3 +1,5 @@
 library indexd_stack_dev;
 
-export 'widget.dart';
+export 'src/controller/lazy_stack_controller.dart';
+export 'src/indexd_animation_type.dart';
+export 'src/widgets/lazy_load_indexed_stack.dart';